+++ CHANGELOG.md
@@ -1,4 +1,14 @@
-# tos-sdk
+# Changelog of @volcengine/tos-sdk
+
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [Unreleased]
+
+### Added
+
+- `getObjectV2` method supports `buffer` in Node.js environment
+- `uploadFile` method supports `dataTransferStatusChange` option
+- `uploadPart` method throws error when there is no etag in browser environment
 
 ## 2.1.1